refactor: Move packages under the pkg namespace
