+++ cmd/avalond/avalond.go
@@ -18,9 +18,9 @@ import (
 	"github.com/0xa1-red/empires-of-avalon/instrumentation/traces"
 	"github.com/0xa1-red/empires-of-avalon/logging"
 	"github.com/0xa1-red/empires-of-avalon/persistence"
-	"github.com/0xa1-red/empires-of-avalon/pkg/auth"
 	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints"
 	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints/registry"
+	"github.com/0xa1-red/empires-of-avalon/pkg/service/auth"
 	"github.com/0xa1-red/empires-of-avalon/protobuf"
 	"github.com/0xa1-red/empires-of-avalon/transport/nats"
 	"github.com/0xa1-red/empires-of-avalon/version"
