+++ http/api/api.go
@@ -1,330 +1,1 @@
 package api
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"log"
-	"net/http"
-
-	"github.com/0xa1-red/empires-of-avalon/http/middleware"
-	"github.com/0xa1-red/empires-of-avalon/instrumentation/traces"
-	"github.com/0xa1-red/empires-of-avalon/pkg/auth"
-	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints"
-	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints/registry"
-	"github.com/0xa1-red/empires-of-avalon/protobuf"
-	"github.com/asynkron/protoactor-go/cluster"
-	jwtmiddleware "github.com/auth0/go-jwt-middleware/v2"
-	"github.com/auth0/go-jwt-middleware/v2/validator"
-	"github.com/go-chi/chi"
-	"github.com/go-chi/cors"
-	"github.com/go-chi/render"
-	"github.com/google/uuid"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/propagation"
-	"golang.org/x/exp/slog"
-	"google.golang.org/protobuf/types/known/timestamppb"
-)
-
-type Router struct {
-	chi.Router
-
-	cluster *cluster.Cluster
-}
-
-func NewRouter(c *cluster.Cluster) *Router {
-	r := chi.NewRouter()
-
-	router := &Router{
-		Router:  r,
-		cluster: c,
-	}
-
-	r.Use(cors.Handler(cors.Options{ // nolint:exhaustruct
-		AllowedOrigins:   []string{"http://localhost:5173"},
-		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
-		AllowCredentials: true,
-		AllowedHeaders:   []string{"Authorization"},
-		MaxAge:           300,
-	}))
-
-	r.Get("/", router.Index)
-
-	r.Group(func(r chi.Router) {
-		r.Use(auth.EnsureValidToken())
-		r.Get("/inventory", router.Inventory)
-		r.Post("/build", router.Build)
-	})
-
-	return router
-}
-
-func (rt *Router) Index(w http.ResponseWriter, r *http.Request) {
-	resp := CommonResponse{
-		Status:     http.StatusOK,
-		StatusText: http.StatusText(http.StatusOK),
-		Message:    "Hi",
-	}
-
-	render.JSON(w, r, resp)
-}
-
-func (rt *Router) Inventory(w http.ResponseWriter, r *http.Request) {
-	ctx, span := traces.Start(r.Context(), "api/router/inventory")
-	defer span.End()
-
-	ctx = context.WithValue(ctx, middleware.RequestIDKey, span.SpanContext().TraceID())
-	r = r.WithContext(ctx)
-
-	w.Header().Set("X-Trace-Id", span.SpanContext().TraceID().String())
-
-	auth := authFromContext(w, r, ctx)
-
-	authUUID, err := uuid.Parse(auth)
-	if err != nil {
-		span.RecordError(err)
-		slog.Error("failed to parse authorization header", err, "auth", auth, "url", r.URL.String())
-
-		status := http.StatusBadRequest
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	span.SetAttributes(attribute.String("user_id", authUUID.String()))
-	slog.Info("getting inventory grain client", "id", blueprints.GetInventoryID(authUUID).String())
-
-	inventory := protobuf.GetInventoryGrainClient(rt.cluster, blueprints.GetInventoryID(authUUID).String())
-
-	carrier := propagation.MapCarrier{}
-	otel.GetTextMapPropagator().Inject(ctx, &carrier)
-
-	res, err := inventory.Describe(&protobuf.DescribeInventoryRequest{
-		TraceID:   carrier.Get("traceparent"),
-		Timestamp: timestamppb.Now(),
-	})
-	if err != nil {
-		span.RecordError(err)
-		slog.Error("failed to get inventory", err, "auth", auth, "url", r.URL.String())
-
-		status := http.StatusInternalServerError
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	render.JSON(w, r, res)
-}
-
-func (rt *Router) Build(w http.ResponseWriter, r *http.Request) {
-	ctx, span := traces.Start(r.Context(), "api/router/build")
-	defer span.End()
-
-	ctx = context.WithValue(ctx, middleware.RequestIDKey, span.SpanContext().TraceID())
-	r = r.WithContext(ctx)
-
-	w.Header().Set("X-Trace-Id", span.SpanContext().TraceID().String())
-
-	auth := authFromContext(w, r, ctx)
-
-	authUUID, err := uuid.Parse(auth)
-	if err != nil {
-		span.RecordError(err)
-		slog.Error("failed to parse authorization header", err, "auth", auth, "url", r.URL.String())
-
-		status := http.StatusBadRequest
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	decoder := json.NewDecoder(r.Body)
-	defer r.Body.Close()
-
-	var buildRequest BuildRequest
-	if err := decoder.Decode(&buildRequest); err != nil {
-		E(w, r, http.StatusBadRequest, err)
-
-		return
-	}
-
-	building := buildRequest.Building
-
-	span.SetAttributes(attribute.String("building", building), attribute.String("user_id", auth))
-
-	amt := getBuildingAmount(buildRequest)
-
-	buildingID := blueprints.GetBuildingID(building)
-
-	log.Println(buildingID)
-
-	blueprint, err := registry.GetBuilding(buildingID)
-	if err != nil {
-		err := fmt.Errorf("invalid building type %s", building)
-		span.RecordError(err)
-		slog.Error("failed to start building", err,
-			"auth", auth,
-			"building", building,
-		)
-
-		status := http.StatusNotFound
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	span.SetAttributes(attribute.String("user_id", authUUID.String()))
-	inventory := protobuf.GetInventoryGrainClient(rt.cluster, blueprints.GetInventoryID(authUUID).String())
-
-	carrier := propagation.MapCarrier{}
-	otel.GetTextMapPropagator().Inject(ctx, &carrier)
-
-	res, err := inventory.Start(&protobuf.StartRequest{
-		TraceID:   carrier.Get("traceparent"),
-		Name:      blueprint.Name.String(),
-		Amount:    amt,
-		Timestamp: timestamppb.Now(),
-	})
-
-	if err != nil {
-		span.RecordError(err)
-		slog.Error("failed to start building", err,
-			"auth", auth,
-			"url", r.URL.String(),
-			"building", blueprint.Name,
-		)
-
-		status := http.StatusInternalServerError
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	if res.Status == protobuf.Status_Error {
-		span.RecordError(err)
-		slog.Error("failed to start building", fmt.Errorf("%s", res.Error),
-			"auth", auth,
-			"url", r.URL.String(),
-			"building", blueprint.Name,
-		)
-
-		status := http.StatusInternalServerError
-		res := ErrorResponse{
-			Status:     status,
-			StatusText: http.StatusText(status),
-			Error:      err,
-		}
-
-		render.Status(r, status)
-		render.JSON(w, r, res)
-
-		return
-	}
-
-	status := http.StatusCreated
-	resp := CommonResponse{
-		Status:     status,
-		StatusText: http.StatusText(status),
-		Message:    "OK",
-	}
-	render.JSON(w, r, resp)
-}
-
-func getBuildingAmount(r BuildRequest) int64 {
-	if !queueBuildings {
-		return 1
-	}
-
-	amt := int64(r.Amount)
-
-	if amt > maximumBuildingRequest {
-		return maximumBuildingRequest
-	}
-
-	return amt
-}
-
-func E(w http.ResponseWriter, r *http.Request, status int, err error) {
-	slog.Error("unauthorized", err, "url", r.URL.String())
-
-	res := ErrorResponse{
-		Status:     status,
-		StatusText: http.StatusText(status),
-		Error:      err,
-	}
-
-	render.Status(r, status)
-	render.JSON(w, r, res)
-}
-
-func authFromContext(w http.ResponseWriter, r *http.Request, ctx context.Context) string {
-	claims := ctx.Value(jwtmiddleware.ContextKey{})
-	if claims == nil {
-		E(w, r, http.StatusInternalServerError, fmt.Errorf("claims not found"))
-		return ""
-	}
-
-	validatedClaims, ok := claims.(*validator.ValidatedClaims)
-	if !ok {
-		E(w, r, http.StatusInternalServerError, fmt.Errorf("failed to validate claims 1"))
-		return ""
-	}
-
-	customClaims, ok := validatedClaims.CustomClaims.(*auth.CustomClaims)
-	if !ok {
-		E(w, r, http.StatusInternalServerError, fmt.Errorf("failed to validate claims 2"))
-		return ""
-	}
-
-	id := customClaims.Subject
-
-	profile, err := auth.GetUserProfile(id)
-	if err != nil {
-		E(w, r, http.StatusInternalServerError, err)
-		return ""
-	}
-
-	if metadata, ok := profile["app_metadata"].(map[string]interface{}); ok {
-		if eid, ok := metadata["external_id"].(string); ok {
-			return eid
-		}
-	}
-
-	return ""
-}
