+++ pkg/service/auth/model.go
