+++ pkg/service/game/config.go
@@ -1,0 +1,6 @@
+package game
+
+var (
+	queueBuildings         bool  = false
+	maximumBuildingRequest int64 = 3
+)
